@@ -37,7 +37,7 @@ project.ext.versionCode = 1_000_000 * major + 10_000 * minor + 100 * patch + bui
 android {
     flavorDimensions "testapp"
 
-    compileSdkVersion 27
+    compileSdkVersion 28
     signingConfigs {
         release {
             keyAlias 'EDRLab'
@@ -49,11 +49,10 @@ android {
     defaultConfig {
         applicationId "org.readium.r2reader"
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode project.ext.versionCode
         versionName project.ext.versionName
-        multiDexEnabled true
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -147,28 +146,28 @@ dependencies {
     }
 
     //    implementation 'androidx.core:core-ktx:0.3'
-    implementation "com.android.support:appcompat-v7:$support_version"
-    implementation "com.android.support:recyclerview-v7:$support_version"
-    implementation "com.android.support:cardview-v7:$support_version"
-    implementation "com.android.support:design:$support_version"
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation "androidx.appcompat:appcompat:1.1.0-alpha02"
+    implementation "androidx.recyclerview:recyclerview:1.1.0-alpha02"
+    implementation "androidx.cardview:cardview:1.0.0"
+    implementation "com.google.android.material:material:1.1.0-alpha03"
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation 'org.nanohttpd:nanohttpd:2.3.2-SNAPSHOT'
     implementation 'org.nanohttpd:nanohttpd-nanolets:2.3.2-SNAPSHOT'
 
     implementation 'com.google.code.gson:gson:2.8.2'
     implementation 'com.squareup.picasso:picasso:2.5.2'
     implementation 'io.reactivex.rxjava2:rxjava:2.1.8'
-    implementation 'org.jetbrains.anko:anko-design:0.10.5'
+    implementation 'org.jetbrains.anko:anko-design:0.10.8'
 
     // For SnackBars
     implementation 'org.jetbrains.anko:anko-commons:0.10.8'
 
     // Anko Layouts
-    implementation 'org.jetbrains.anko:anko-sdk25:0.10.5'
+    implementation 'org.jetbrains.anko:anko-sdk25:0.10.8'
 
     // sdk15, sdk19, sdk21, sdk23 are also available
-    implementation 'org.jetbrains.anko:anko-appcompat-v7:0.10.5'
-    implementation 'org.jetbrains.anko:anko-recyclerview-v7:0.10.5'
+    implementation 'org.jetbrains.anko:anko-appcompat-v7:0.10.8'
+    implementation 'org.jetbrains.anko:anko-recyclerview-v7:0.10.8'
 
     // Coroutine listeners for Anko Layouts
     implementation 'org.jetbrains.anko:anko-sdk25-coroutines:0.10.8'
@@ -208,7 +207,7 @@ dependencies {
     implementation 'org.jsoup:jsoup:1.11.2'
 
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test:runner:1.1.2-alpha01'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.2-alpha01'
     testImplementation "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
 }
