@@ -11,9 +11,9 @@
 package org.readium.r2.testapp
 
 import android.content.Context
-import android.support.annotation.ColorInt
-import android.support.annotation.ColorRes
-import android.support.v4.content.ContextCompat
+import androidx.annotation.ColorInt
+import androidx.annotation.ColorRes
+import androidx.core.content.ContextCompat
 import java.io.File
 import java.io.FileFilter
 import java.io.InputStream
