@@ -12,8 +12,8 @@ package org.readium.r2.testapp.permissions
 
 import android.app.Activity
 import android.content.pm.PackageManager
-import android.support.v4.app.ActivityCompat
-import android.support.v4.content.ContextCompat
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
 import io.reactivex.Single
 import io.reactivex.subjects.PublishSubject
 
